Extract base class for manip interface decorators
